feat: Redis 설정 #93
- RootConfig의 Redis 설정 RedisConfig로 이동
- 상환 시뮬레이션 결과에 맞는 템플릿 설정
- Spring Data Redis 버전 변경
